Update Coinelity (Middleware & Backend)
- Changed the names from Binary Options to just options;
- Refactored the ApiResponse no ease success/error message with only one message;
- Other minor refactorings.

+++ Coinelity.AspServer/Models/DTO's/ClosedOptionDTO.cs
@@ -56,7 +56,7 @@ namespace Coinelity.AspServer.Models
 
         public int LifetimeLabel { get; set; }
 
-        public sbyte PayoutPercent { get; set; }
+        public short PayoutPercent { get; set; }
 
         public decimal StrikePrice { get; set; }
